Update deployment.amd64.json
I changed the Github repo from packages to ghcr.io. This is the new Repo for docker containers on Github that we need to use. I also changes the build action and used a script that was developed for ww.Bluegit.com docker build actions and the basis. We can now create containers and deploy them to an Edge device.